Added print Last 5

+++ Software/main.py
@@ -1,35 +1,65 @@
 
 from tinydb import TinyDB, Query
+from os import system, name
 import time
+
+
+def clear():
+    if (name == 'nt'):
+        system('cls')
+    # for mac and linux
+    else:
+        system('clear')
+
+
+def printLast5():
+    Working = Contact()
+    for idoc_id in range(len(db) - 4, len(db) + 1):
+        Working.load(db.get(doc_id=idoc_id))
+        Working.print()
 
 
 class Contact:
     def __init__(self):
         self.time = 0
-        self.transmiting = ""
+        self.transmitting = ""
         self.receiving = ""
         self.message = ""
 
     def log(self, db):
-        db.insert({'time': self.time, 'Tx': self.transmiting,
+        db.insert({'time': self.time, 'Tx': self.transmitting,
                    'Rx': self.receiving, 'msg': self.message})
 
     def getContactInfo(self):
-        self.transmiting = input("Station: ")
+        self.transmitting = input("Station: ")
         self.time = time.time()
         self.message = input("Message: ")
 
+    def print(self):
+        print(
+            f"Rx: {self.receiving}  Tx: {self.transmitting} Time: {time.ctime(self.time)}   Message: {self.message}")
 
+    def load(self, doc):
+        self.time = doc['time']
+        self.transmitting = doc['Tx']
+        self.receiving = doc['Rx']
+        self.message = doc['msg']
+
+
+# Init
 db = TinyDB('./test.json')
-db.purge()
-contact = Contact()
 User = Query()
-contact.receiving = "NC"
-while (input("Continue:").lower() == "y"):
-    contact.getContactInfo()
-    contact.log(db)
+WorkingContact = Contact()
+clear()
 
-last = db.get(doc_id=len(db))
-print("Last contact:")
-print(
-    f"Rx: {last['Rx']}\nTx: {last['Tx']}\nTime: {time.ctime(last['time'])}\nMessage: {last['msg']}")
+# Initial Config
+print("Welcome to Comms Log by Thomas Fike KG7FXT")
+if (input("Do you want to start a new document? [Y/n]: ").lower() == "y"):
+    db.purge()
+Rx = input("What would you like to call your Receiving Station: ")
+WorkingContact.receiving = Rx
+# for i in range(0, 7):
+#    WorkingContact.getContactInfo()
+#    WorkingContact.log(db)
+
+printLast5()
